add helper function
